docs(example): update example app with delegate
